pyproject.toml
[tool.poetry]
name = "movie-bet-bot"
version = "0.1.0"
description = "Discord bot that tracks multiple Letterboxd lists and periodically updates a channel with their status."
authors = ["Luke Schmitt <luke@lukeschmitt.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "movie_bet_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
python-dotenv = "^0.21.0"
beautifulsoup4 = "^4.11.1"
pyyaml = "^6.0"
tmdbsimple = "^2.9.1"
html2image = "^2.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
lark = "^1.1.5"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
